+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - 🟥 rename `is.set` to `is.happy`
 - 🟥 rename `is.unset` to `is.sad`
 - 🟥 pubsub changes
-  - 🟥 change `xub`, now returns non-invocable full pubsub facilities (used to return a tuple)
+  - 🟥 change `xub`, pubsub facilities object (used to return a tuple)
   - 🍏 add new `.publish` which is a normal publisher fn without the xub facilities
   - 🍏 add new `.subscribe` which is a normal subscriber fn without the xub facilities
   - 🍏 add `fn` to `.next(fn)`